+++ example/next/babel.config.js
@@ -1,9 +1,9 @@
 const path = require('path');
 module.exports = {
-  presets: ['@expo/next-adapter/babel'],
+  presets: ['next/babel'],
   plugins: [
-    ['@babel/plugin-proposal-private-methods', { loose: true }],
-    ['@babel/plugin-proposal-private-property-in-object', { loose: true }],
+    // ['@babel/plugin-proposal-private-methods'],
+    // ['@babel/plugin-proposal-private-property-in-object'],
     process.env.NODE_ENV !== 'production'
       ? [
           'module-resolver',