+++ packages/react/src/resolver/StyledValueToCSSObject.ts
@@ -1,0 +1,13 @@
+import type { CSSObject, StyledValue } from '../types';
+import { resolvedTokenization } from '../utils';
+
+export function StyledValueToCSSObject(
+  input: StyledValue | undefined,
+  CONFIG: any
+): CSSObject {
+  if (!input) {
+    return {};
+  }
+  // return input;
+  return resolvedTokenization(input, CONFIG);
+}