+++ packages/react/src/utils.ts
@@ -85,6 +85,11 @@ export function resolveStringToken(
           modifiedTokenScale = 'space';
         }
 
+        if (!config || !config.tokens) {
+          throw new Error(
+            'You cannot use tokens without wrapping the component with StyledProvider. Please wrap the component with a StyledProvider and pass theme config.'
+          );
+        }
         if (
           config?.tokens[modifiedTokenScale] &&
           config?.tokens[modifiedTokenScale].hasOwnProperty(